requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.1
Flask-Login==0.6.3
python-dotenv==1.0.0
pymysql==1.1.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
openpyxl==3.1.2
playwright==1.40.0
selenium==4.16.0
webdriver-manager==4.0.1
lxml==4.9.3
urllib3==2.2.0
chardet==5.2.0
charset-normalizer==3.3.2
celery==5.3.4
redis==5.0.1
validators==0.22.0
undetected-chromedriver==3.5.5
dnspython>=2.4.0

# AI-powered extraction
crawl4ai>=0.4.0                  # Intelligent web crawling with JS rendering
trafilatura>=2.0.0               # Content extraction from HTML
html-text>=0.6.0                 # HTML to text conversion
langextract>=1.1.1               # PRIMARY: Structured extraction via Ollama (local, no keys required)
onprem>=0.6.0                    # Optional: For local .gguf model inference via llama-cpp-python
                                 # If using onprem for local models, also install: pip install llama-cpp-python
sentence-transformers>=3.0.0     # Embeddings for semantic search
haystack-ai>=2.6.0               # RAG & retrieval pipeline
numpy>=1.24.0                    # Numerical computations
